add DataFeature, refactor to update other things, HistogramGraphic is broken because of DataFeature
