+++ fastplotlib/graphics/features/__init__.py
@@ -1,0 +1,4 @@
+from ._colors import ColorFeature
+from ._data import DataFeature
+from ._present import PresentFeature
+from ._base import GraphicFeature
